Add standalone demo.
This demo uses the raw files from Andrey Pshenov's PhD thesis
simulations, converted to SOLPS 5 format by Vlad Neverov. The demo
contains similar functionality to the existing AUG/raw, MDSplus and
balance.nc demos. Additionally, it demonstrates how to add emission
models to the plasma and image the plasma with a visible camera,
providing a complete workflow example which should be useful to new
Cherab users and familiar to experienced users new to the SOLPS
package.