Add more general parser for fort.44 files which include labels
Versions of fort.44 >= 20160829 have a comment line before each data
block which includes the name of the field and the size of the data
block. This makes the file close enough to self-describing that the
block can be parsed without knowledge of the exact format. This should
enable greater flexibility for the number of formats we support.

This change means we essentially have two types of parsers, unlabelled
and labelled. The unlabelled ones are further split into pre-2007 and
2013 at the moment, due to non-self-describing differences in these
formats. But the availability of at least a couple of general parsers
leads to a reworking of the logic to choose a parser: it's no longer
tied to the exact version but is more of a boundary match.