Add more general parser for fort.44 files which include labels (#41)
* Add more general parser for fort.44 files which include labels

Versions of fort.44 >= 20160829 have a comment line before each data
block which includes the name of the field and the size of the data
block. This makes the file close enough to self-describing that the
block can be parsed without knowledge of the exact format. This should
enable greater flexibility for the number of formats we support.

This change means we essentially have two types of parsers, unlabelled
and labelled. The unlabelled ones are further split into pre-2007 and
2013 at the moment, due to non-self-describing differences in these
formats. But the availability of at least a couple of general parsers
leads to a reworking of the logic to choose a parser: it's no longer
tied to the exact version but is more of a boundary match.

* Close fort.44 file as soon as reading is finished

No need to keep the file open when setting Eirene properties.

* Make molecules and ions optional when reading labelled fort.44s

* Return a dictionary tuple from the labelled fort44 block parser

This is more appropriate than a (name, value) tuple, particularly as
it's being added to a dictionary anyway.

* Deal with rogue blank lines in the labelled fort44 parser

Although blank lines shouldn't exist, fort.44 files have been found in
the wild where they do. The check for EOF has been refined so it isn't
triggered by a blank line, and blank lines are treated as null data by
the block reader.

* Cope with non-species-resolved radiated power in raw Eirene output

Some EIRENE/SOLPS combinations output eneutrad etc. summed over all
species, rather than resolved over species. This is easy to detect and
correct for, since we always output the total radiated power summed
over all species anyway.