Fix loading of total radiated power from raw files (#42)
Load Eirene total radiated power in W/m3, where positive values
correspond to emission of radiation. Add line emission and
Bremsstrahlung from the plasma solution using the b2fplasmf file:
without this a significant fraction of the radiation is omitted. This
does mean a b2fplasmf file is required: if this is not available then
no value is assigned to SOLPSSimulation.total_radiation, as using the
Eirene neutral data alone would be misleading.

With this change, the SOLPSSimulation.total_radiation quantity is
comparable for simulations loaded from raw files, MDSplus or a
balance.nc file.

Fixes #28